requirements.txt
sentence-transformers==2.6.1
scikit-learn==1.3.2
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.2
nltk==3.8.1
streamlit==1.32.2
pandas==2.2.0
